requirements.txt
Django==4
pyyaml

